+++ src/main/java/smida/test_task/havriushenko/stockRegestry/models/StockModel.java
@@ -17,13 +17,12 @@ public class StockModel {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private long pk;
-
     @Column(length = 2000)
     private String comment;
     @Column(name = "usreou")
     private long USREOU;            //Unified State Register of Enterprises and Organizations of Ukraine
     @Column
-    private double quantity;
+    private long quantity;
     @Column(name = "total_nominal_value")
     private double totalNominalValue;
     @Column(name = "nominal_value")
